Final
Note that this experiment is supposed to be run on a touch screen.

Note that the tone is now presented centrally instead of indicating on what side the target will appear. This was done to directly replicate past gap effect studies (it was my mistake before to have the tone presented on one side or the other).

Note that there are 30 practice trials and 480 actual trials. You can quit by pressing 'q'.